+++ Changes.md
@@ -10,6 +10,7 @@ Breaking Changes
   - Removed compatibility shim for referencing `GafferCortex` attributes via the `Gaffer` namespace.
   - Removed compatibility shim for referencing `GafferDispatch` attributes via the `Gaffer` namespace.
   - Removed compatibility shim for referencing `MetadataAlgo` and `MonitorAlgo` without namespace qualification.
+  - Removed compatibility shim for referencing `UndoScope` as `UndoContext`.
 - GafferUI : Removed compatibility shim for referencing `GafferCortexUI` attributes via the `GafferUI` namespace.
 - GafferImage :
   - Removed compatibility for loading old `RemoveChannels` nodes.